+++ changelog.md
@@ -2,6 +2,8 @@
 
 //  Add your own contribution below
 
+* Add Jenkins CI source - macklinu
+
 ### 0.7.0